+++ discord/objects/components/__init__.py
@@ -1,5 +1,6 @@
 from .checkbox import CheckboxComponent
 from .checkbox_group import CheckboxGroupComponent
 from .checkbox_group_option import CheckboxGroupOption
+from .file_upload import FileUploadComponent
 from .radio_group import RadioGroupComponent
 from .radio_group_option import RadioGroupOption