+++ docs/source/changelog.rst
@@ -38,6 +38,7 @@ v26.1.12-dev
 - Added :class:`~discord.objects.components.CheckboxGroupOption` object.
 - Added :class:`~discord.objects.components.RadioGroupComponent` component.
 - Added :class:`~discord.objects.components.RadioGroupOption` object.
+- Added :class:`~discord.objects.components.FileUploadComponent` component.
 
 **Documentation Changes**